materialcache and refcountedid
